Adding a way to specify "not to run" tests
Some tests have the potential of causing system crashes / hangs.
Such test in cfg file hinders the run of tests that follow. So,
if such tests are identified and need to be "not run" for a
particular environment (Since they can run fine on others), we
have a provision to mention such tests to not run.

Add tests that you dont want to run as comma separated list, in the
particular environment, in the file config/wrapper/no_run_tests.conf.
The tests should be taken from cfg files.
The mentioned tests, if present in any cfg file, do not run.

Each test-yaml pair is treated as a different test, and test without
yaml is treated as a different test, as should be.

Signed-off-by: Narasimhan V <sim@linux.vnet.ibm.com>