Fix for norun_tests not found in an environment
We get an error in case there is no norun_<env> found in
norun_tests conf.
This is fixed by handling it when not available in the wrapper.

We actually had discovered this error since there was a refactor
in the environment, and had missed norun_tests in the process.

So, this patch takes care of fixing it in both the places.

Signed-off-by: Narasimhan V <sim@linux.vnet.ibm.com>
Reported-by: Abdul Haleem <abdhalee@linux.vnet.ibm.com>