Refactor package selection based on major version
This patch provides way to select packages based on major version
of a distribution (which does not change often). It also falls
back to the package with base name if the running major version is
not handled.

i.e
rhel7.6 & rhel7.7
rhel8.0 & rhel8.1
sles15 & sles15.sp1

All pair use similar list of packages.

When run on a "rhel6" environment it falls back to "rhel"

As for the norun configuration, it uses the combination of major
and minor version number so that it is precise to a current test
release target.

Signed-off-by: Harish <harish@linux.vnet.ibm.com>