Fixing result summary for not existing config file
Result summary beautification was done, which had not considered
non existing config files or tests which had not run, or job.log
not generated for some reason.

We find the length needed for column 1, which is the longest of:

    TestSuiteName
    job.log path
    result link

By considering all 3 of these from all testsuites, wherever
applicable, we are fixing that case.

Signed-off-by: Narasimhan V <sim@linux.vnet.ibm.com>
Reported-by: Satheesh Rajendran <sathnaga@linux.vnet.ibm.com>