Add support for new avocado config value: optimize
Adding support for new config value:
sysinfo.collect.optimize

Setting this parameter to True will not log any unchanged sysinfo
in the `post` section of the test / job sysinfo folder. It checks
for `pre` section to find if they are changed or unchanged.
Setting it to False will log all sysinfo, as defined via the
avocado config. This is the default avocado behaviour.

Via this patch, we are setting the default behaviour when used
wrapper, to `True`.

References:
RFC: avocado-framework/avocado/issues/3843
Implementation: avocado-framework/avocado/pull/3924

Signed-off-by: Narasimhan V <sim@linux.vnet.ibm.com>