Generate .html analysis file for avocado runs (#329)
In addition to Analysis.xlsx, this patch generates
Analysis.html file as output. This can be generated
for new analysis as well as comparison analysis.

The Analysis.xlsx file is converted to Analysis.html
file and both the outputs are generated. This .html
file can be used to display the summary of the avocado
based results in a web page for a dashboard based view.

This .html output has filter table drop downs as an
enhancement to the .xlsx output. This filter can be used
for further breaking down of the results.

Signed-off-by: Misbah Anjum N <misanjum@linux.vnet.ibm.com>